pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kgentik"
version = "0.1.0"
description = "Kgentik SDK for seamless tool resolution and agent development"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Kgentik Team"},
]
keywords = ["ai", "agents", "tools", "langgraph", "langchain"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "isort>=5.12",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/kgentik/kgentik"
Documentation = "https://docs.kgentik.com"
Repository = "https://github.com/kgentik/kgentik"
Issues = "https://github.com/kgentik/kgentik/issues"

[tool.hatch.build.targets.wheel]
packages = ["kgentik"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88
